Persisting commands instead of events which make more sense

+++ src/main/scala/nl/rubenernst/mrdeveloper/scala/protocol/RepositoryProtocol.scala
@@ -9,14 +9,14 @@ object RepositoryProtocol {
   }
 
   final case class NewRepository(id: Long, fullName: String, description: Option[String]) extends RepositoryCommand
-  final case class StarRepository(id: Long, fullName: String, stars: Int) extends RepositoryCommand
+  final case class StarRepository(id: Long, fullName: String, stars: Long) extends RepositoryCommand
 
   sealed trait RepositoryEvent {
     def id: Long
   }
 
   final case class RepositoryCreated(id: Long, fullName: String, description: Option[String]) extends RepositoryEvent
-  final case class RepositoryStarred(id: Long, fullName: String, stars: Int) extends RepositoryEvent
+  final case class RepositoryStarred(id: Long, fullName: String, stars: Long) extends RepositoryEvent
 
   final case class ErrorMessage(data: String)
 }
